add authorization

+++ controllers/feed.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
-
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1
@@ -51,15 +51,25 @@ exports.createPost = (req, res, next) => {
   const post = new Post({
     title: title,
     content: content,
-    creator: { name: 'cz' },
+    creator: req.userId,
     imageUrl: imageUrl
   });
 
+  let creator;
   post.save()
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      creator.posts.push(post);
+      return creator.save();
+    })
     .then(result => {
       res.status(201).json({
         message: 'Create post successfully!',
-        post: result
+        post: post,
+        creator: {_id: creator._id, name: creator.name}
       });
     })
     .catch(err => {
@@ -102,9 +112,21 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if(post.creator != req.userId) {
+        const error = new Error('Not Authorized.')
+        error.statusCode = 403;
+        throw error;
+      }
       // check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
+    })
+    .then(result => {
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then(result => {
       res.status(200).json({ message: 'Post deleted.' });
